backend/scrapers/lidl_scraper.py
from typing import List, Dict
import httpx
from bs4 import BeautifulSoup
from datetime import datetime, timedelta
from .base_scraper import BaseScraper


class LidlScraper(BaseScraper):
    """Scraper for Lidl deals"""

    def __init__(self):
        super().__init__(
            store_name='Lidl',
            base_url='https://www.lidl.de'
        )

    async def scrape_deals(self) -> List[Dict]:
        """
        Scrape deals from Lidl website

        Note: This is a template implementation.
        Lidl's actual website structure may require Playwright/Selenium
        for JavaScript-rendered content.
        """
        deals = []

        try:
            # Lidl Prospekt-Seite
            url = f'{self.base_url}/c/angebote-diese-woche/a10006065'

            async with httpx.AsyncClient(timeout=30.0) as client:
                response = await client.get(url, follow_redirects=True)

                if response.status_code != 200:
                    self.logger.error(f"Failed to fetch Lidl page: {response.status_code}")
                    return []

                soup = BeautifulSoup(response.text, 'html.parser')

                # Find product cards (structure may vary)
                # This is a template - actual selectors need to be adjusted
                product_cards = soup.find_all('div', class_='product')

                for card in product_cards[:50]:  # Limit to 50 deals
                    try:
                        # Extract product info (adjust selectors based on actual HTML)
                        name_elem = card.find('h3') or card.find('div', class_='name')
                        price_elem = card.find('span', class_='price')
                        image_elem = card.find('img')

                        if not name_elem or not price_elem:
                            continue

                        product_name = name_elem.get_text(strip=True)
                        price_text = price_elem.get_text(strip=True)
                        discount_price = self._parse_price(price_text)

                        if not discount_price:
                            continue

                        image_url = None
                        if image_elem:
                            image_url = image_elem.get('src') or image_elem.get('data-src')
                            if image_url and not image_url.startswith('http'):
                                image_url = self.base_url + image_url

                        # Categorize product
                        category = self._categorize_product(product_name)

                        deal = self._create_deal(
                            product_name=product_name,
                            discount_price=discount_price,
                            image_url=image_url,
                            category=category,
                            valid_from=datetime.now(),
                            valid_until=datetime.now() + timedelta(days=7)
                        )

                        deals.append(deal)

                    except Exception as e:
                        self.logger.warning(f"Error parsing Lidl product: {e}")
                        continue

            self.logger.info(f"Scraped {len(deals)} deals from Lidl")
            return deals

        except Exception as e:
            self.logger.error(f"Error scraping Lidl: {e}")
            return []

    def _categorize_product(self, product_name: str) -> str:
        """Categorize product based on name"""
        name_lower = product_name.lower()

        if any(word in name_lower for word in ['fleisch', 'hähnchen', 'rind', 'schwein', 'wurst']):
            return 'Fleisch'
        elif any(word in name_lower for word in ['gemüse', 'salat', 'tomate', 'gurke', 'paprika', 'brokkoli']):
            return 'Gemüse'
        elif any(word in name_lower for word in ['obst', 'apfel', 'banane', 'orange', 'beeren']):
            return 'Obst'
        elif any(word in name_lower for word in ['milch', 'käse', 'joghurt', 'butter', 'sahne']):
            return 'Milchprodukte'
        elif any(word in name_lower for word in ['fisch', 'lachs', 'thunfisch']):
            return 'Fisch'
        elif any(word in name_lower for word in ['brot', 'brötchen', 'kuchen']):
            return 'Backwaren'
        elif any(word in name_lower for word in ['getränk', 'saft', 'wasser', 'cola', 'bier', 'wein']):
            return 'Getränke'
        else:
            return 'Grundnahrungsmittel'
