backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
beautifulsoup4>=4.12.0
requests>=2.31.0
httpx>=0.24.0
python-dateutil>=2.8.0

# File upload & processing
python-multipart>=0.0.6
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# AI/ML for deal extraction
openai>=1.0.0
